requirements.txt
streamlit
opencv-python
numpy
ultralytics
pillow
transformers
langchain
huggingface_hub
